升级.net10，升级Nuget依赖，调整项目依赖顺序

+++ api/PurestAdmin.SqlSugar/Entity/RequestLogEntity.cs
@@ -32,7 +32,7 @@ public partial class RequestLogEntity : BaseEntity
     /// 执行耗时
     /// </summary>
     [SugarColumn(ColumnName = "ELAPSED_TIME")]
-    public long ElapsedTime { get; set; }
+    public int ElapsedTime { get; set; }
     /// <summary>
     /// 客户端IP
     /// </summary>
